requirements_base.txt
otree-core==0.4.34
Django==1.8.8
